Batch combiner bug fixes. Detect attempts to add a batch to a group/frame twice in debug builds.

+++ Squared/RenderLib/Convenience.cs
@@ -98,7 +98,7 @@ namespace Squared.Render.Convenience {
         public readonly bool UseZBuffer;
         public readonly bool AutoIncrementSortKey;
 
-        private int NextSortKey;
+        private float NextSortKey;
 
         public BitmapRenderer (
             IBatchContainer container,
